// openapi/models/update-users-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UpdateUsersRequest
 */
export interface UpdateUsersRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateUsersRequest
     */
    'firstname': string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUsersRequest
     */
    'lastname': string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUsersRequest
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUsersRequest
     */
    'invitationcode'?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateUsersRequest
     */
    'tenantid'?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateUsersRequest
     */
    'status'?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateUsersRequest
     */
    'id': number;
}